made seperate activity for reading data
