+++ app/src/main/AndroidManifest.xml
@@ -34,17 +34,21 @@
             android:exported="false" />
         <activity
             android:name=".healthConnect.HealthConnectActivity"
-            android:exported="true"
-            android:enabled="true">
+            android:enabled="true"
+            android:exported="true">
+
             <!-- TODO: Required to specify which Health Connect permissions the app can request -->
             <meta-data
                 android:name="health_permissions"
-                android:resource="@array/health_permissions"/>
+                android:resource="@array/health_permissions" />
             <!-- (b) handle intent -->
             <intent-filter>
                 <action android:name="androidx.health.ACTION_SHOW_PERMISSIONS_RATIONALE" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".healthConnect.ReadDataActivity"
+            android:exported="false" />
         <activity
             android:name=".fitbit.FitbitActivity"
             android:exported="false" />
