+++ app/src/main/java/com/devsoc/hrmaa/MainActivity.kt
@@ -3,8 +3,10 @@ package com.devsoc.hrmaa
 
 import android.content.Intent
 import android.os.Bundle
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
+import androidx.health.connect.client.HealthConnectClient
 import com.devsoc.hrmaa.bluetooth.AvailableDevicesActivity
 import com.devsoc.hrmaa.databinding.ActivityMainBinding
 import com.devsoc.hrmaa.ecg.ECGActivity
@@ -20,17 +22,26 @@ class MainActivity : AppCompatActivity() {
         binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
 
 
-        binding.hcCvMa.setOnClickListener{
-            startActivity(Intent(this,HealthConnectActivity::class.java))
+        binding.hcCvMa.setOnClickListener {
+            //launch only if Health Connect is installed on device
+            if (HealthConnectClient.isAvailable(this)) {
+                startActivity(Intent(this, HealthConnectActivity::class.java))
+            } else {
+                Toast.makeText(
+                    this,
+                    "Please install the Google Health Connect App first.",
+                    Toast.LENGTH_SHORT
+                ).show()
+            }
         }
-        binding.fitbitCvMa.setOnClickListener{
-            startActivity(Intent(this,FitbitActivity::class.java))
+        binding.fitbitCvMa.setOnClickListener {
+            startActivity(Intent(this, FitbitActivity::class.java))
         }
-        binding.ppgCvMa.setOnClickListener{
+        binding.ppgCvMa.setOnClickListener {
             startActivity(Intent(this, PPGActivity::class.java))
         }
-        binding.ecgCvMa.setOnClickListener{
-            startActivity(Intent(this,AvailableDevicesActivity::class.java))
+        binding.ecgCvMa.setOnClickListener {
+            startActivity(Intent(this, AvailableDevicesActivity::class.java))
         }
