+++ app/src/main/res/layout/activity_health_connect.xml
@@ -12,15 +12,27 @@
         android:layout_height="match_parent"
         tools:context=".healthConnect.HealthConnectActivity">
 
-        <TextView
-            android:id="@+id/data_tv_hca"
+        <androidx.cardview.widget.CardView
+            android:id="@+id/read_cv_hca"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:textSize="20sp"
+            app:cardCornerRadius="10dp"
+            app:cardElevation="5dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"></TextView>
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintVertical_bias="0.25">
 
+            <TextView
+                android:id="@+id/read_tv_hca"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_margin="20dp"
+                android:fontFamily="@font/orbitron_regular"
+                android:text="Read Data"
+                android:textSize="25sp"
+                android:textStyle="bold" />
+        </androidx.cardview.widget.CardView>
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
