+++ app/src/main/res/values/strings.xml
@@ -38,4 +38,6 @@
         <item quantity="one">Pulse: %2.1f (%d cycle in %2.1f seconds)</item>
         <item quantity="other">Pulse: %2.1f (%d cycles in %2.1f seconds)</item>
     </plurals>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
